libs/messaging/pyproject.toml
[project]
name = "cogniverse-messaging"
version = "0.1.0"
description = "Messaging gateway for Telegram/Slack integration"
requires-python = ">=3.12"
dependencies = [
    "python-telegram-bot>=20.0",
    "httpx>=0.27",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["cogniverse_messaging"]
